+++ api/firebase_config.py
@@ -1,5 +1,5 @@
 import pyrebase
-from api.config import config
+from config import config
 
 firebase = pyrebase.initialize_app(config)
 auth = firebase.auth()
